fix(sessions): chunked select-all, trim ids, refetch all, split row
- bridge clamp 5000 + toggleSelectAll chunk 2000 loop fixes >2000 truncation
- session delete/restore/reveal trim ids before validate/join/index + fs_guard block \n\r
- areAll/isIndeterminate precise (size===total + hasNextPage/size<total vs items checks)
- hooks invalidate refetchType all for infinite/paged + legacy
- extract duplicate preview to truncatePreview/buildPreviewCapped, i18n sessions.stat.turns/steps, aria-label title fallback
- split 534-line config-sessions: extract SessionRow to session-row.tsx (534->402 lines), keep helpers
- config-dialog Switch order align with navs (sessions after profiles)
