+++ src/components/config-sessions.tsx
@@ -1,5 +1,5 @@
-import { ArrowRotateRight, FolderOpen, TrashBin } from '@gravity-ui/icons'
-import { Button, Checkbox, Chip, Input, Label, Spinner } from '@heroui/react'
+import { ArrowRotateRight, TrashBin } from '@gravity-ui/icons'
+import { Button, Checkbox, Chip, Input, Spinner } from '@heroui/react'
 import { useOverlay } from '@overlastic/react'
 import { useVirtualizer } from '@tanstack/react-virtual'
 import { invoke } from '@tauri-apps/api/core'
@@ -9,146 +9,28 @@ import { If } from 'react-if-lite'
 import { useDshSessionsInfinite } from '@/hooks/use-dsh-sessions-paged'
 import type { SessionFileInfo } from '@/hooks/use-dsh-sessions'
 import { toast } from '@/utils/toast'
-import { Ellipsis } from './ellipsis'
 import { Empty } from './empty'
-import { Item } from './item'
 import { Modal } from './modal'
 import { PanelHeader } from './panel-header'
 import { PanelState } from './panel-state'
+import { SessionRow } from './session-row'
 
 type FilterType = 'all' | 'active' | 'archived' | 'orphan'
 type SortKey = 'createdAt' | 'size' | 'turns'
-function formatSize(bytes: number) {
-  if (bytes < 1024) return `${bytes} B`
-  if (bytes < 1024 * 1024) return `${(bytes / 1024).toFixed(1)} KB`
-  if (bytes < 1024 * 1024 * 1024) return `${(bytes / 1024 / 1024).toFixed(1)} MB`
-  return `${(bytes / 1024 / 1024 / 1024).toFixed(2)} GB`
+
+function truncatePreview(text: string, max = 24) {
+  const tt = text.trim()
+  return tt.length > max ? `${tt.slice(0, max)}…` : tt
 }
 
-function formatTime(ts: number) {
-  if (!ts) return '-'
-  const d = new Date(ts)
-  if (Number.isNaN(d.getTime())) return '-'
-  return d.toLocaleString()
+function buildPreviewCapped(ids: string[], items: SessionFileInfo[], t: (k: string, o?: unknown) => string) {
+  const raw = ids.slice(0, 3).map(id => truncatePreview(items.find(s => s.id === id)?.title || id))
+  const preview = raw.join('、')
+  const more = ids.length > 3 ? t('sessions.delete.batch_more', { count: ids.length - 3 }) as string : ''
+  const text = preview + more
+  return text.length > 72 ? `${text.slice(0, 72)}…` : text
 }
 
-function statusChipColor(status: SessionFileInfo['archivedStatus']) {
-  if (status === 'active') return 'success' as const
-  if (status === 'archived') return 'warning' as const
-  return 'default' as const
-}
-
-interface SessionRowProps {
-  session: SessionFileInfo
-  selected: boolean
-  onToggle: () => void
-  onDelete: () => void
-  onRestore: () => void
-  onOpenDir: () => void
-  deletePending: boolean
-  restorePending: boolean
-  isOpening: boolean
-}
-
-function SessionRow({ session: s, selected, onToggle, onDelete, onRestore, onOpenDir, deletePending, restorePending, isOpening }: SessionRowProps) {
-  const { t } = useTranslation()
-  return (
-    <Item
-      left={(
-        <div className="min-w-0 flex flex-col gap-1">
-          <div className="flex min-w-0 items-center gap-1.5">
-            <Checkbox
-              isSelected={selected}
-              onChange={onToggle}
-              isDisabled={deletePending}
-              aria-label={s.id}
-              className="shrink-0"
-            >
-              <Checkbox.Content>
-                <Checkbox.Control>
-                  <Checkbox.Indicator />
-                </Checkbox.Control>
-              </Checkbox.Content>
-            </Checkbox>
-            <Label className="min-w-0 truncate text-sm font-medium text-ink">
-              {s.title || t('sessions.untitled')}
-            </Label>
-            <Chip size="sm" variant="soft" color={statusChipColor(s.archivedStatus)} className="shrink-0 rounded-md font-medium">
-              {t(`sessions.status.${s.archivedStatus}`)}
-            </Chip>
-            <If cond={s.isEmpty}>
-              <Chip size="sm" variant="soft" color="warning" className="shrink-0 rounded-md font-medium">
-                {t('sessions.empty')}
-              </Chip>
-            </If>
-            <If cond={!!s.isParseFailed}>
-              <Chip size="sm" variant="soft" color="danger" className="shrink-0 rounded-md font-medium">
-                {t('sessions.parse_failed')}
-              </Chip>
-            </If>
-          </div>
-          <div className="ml-7 flex flex-wrap items-center gap-x-3 gap-y-1 text-xs text-muted">
-            <span>{formatSize(s.size)}</span>
-            <span>
-              turns
-              {' '}
-              {s.turns}
-            </span>
-            <span>
-              steps
-              {' '}
-              {s.steps}
-            </span>
-            <span>{formatTime(s.createdAt)}</span>
-          </div>
-          <Ellipsis className="ml-7 text-xs text-muted">
-            {s.cwd || s.id}
-          </Ellipsis>
-        </div>
-      )}
-      right={(
-        <>
-          <If cond={s.archivedStatus === 'archived'}>
-            <Button
-              size="sm"
-              variant="tertiary"
-              className="h-6 w-6 shrink-0 rounded-md p-0"
-              onPress={onRestore}
-              isDisabled={restorePending || deletePending || isOpening}
-              aria-label={t('sessions.restore.one')}
-            >
-              <If cond={restorePending} else={<ArrowRotateRight className="size-3.5 scale-x-[-1]" />}>
-                <Spinner size="sm" color="current" />
-              </If>
-            </Button>
-          </If>
-          <Button
-            size="sm"
-            variant="tertiary"
-            className="h-6 w-6 shrink-0 rounded-md p-0"
-            onPress={onOpenDir}
-            isDisabled={isOpening || deletePending || restorePending}
-            aria-label={t('sessions.open_dir')}
-          >
-            <If cond={isOpening} else={<FolderOpen className="size-3.5" />}>
-              <Spinner size="sm" color="current" />
-            </If>
-          </Button>
-          <Button
-            size="sm"
-            variant="tertiary"
-            className="h-6 w-6 shrink-0 rounded-md p-0"
-            onPress={onDelete}
-            isDisabled={deletePending || isOpening || restorePending}
-            aria-label={t('sessions.delete.one')}
-          >
-            <TrashBin className="size-3.5" />
-          </Button>
-        </>
-      )}
-    />
-  )
-}
 
 export function ConfigSessions() {
   const { t } = useTranslation()
@@ -172,9 +54,8 @@ export function ConfigSessions() {
     sortKey,
     sortAsc,
   })
-  const areAllFilteredSelected = total > 0 && selected.size === total && (hasNextPage ? true : items.length > 0 && items.every(s => selected.has(s.id)))
-  const isIndeterminate = !areAllFilteredSelected && selected.size > 0 && (hasNextPage ? true : items.some(s => selected.has(s.id)))
-
+  const areAllFilteredSelected = total > 0 && selected.size === total && (hasNextPage ? true : items.length === total && items.length > 0 && items.every(s => selected.has(s.id)))
+  const isIndeterminate = !areAllFilteredSelected && selected.size > 0 && (hasNextPage ? selected.size < total : items.some(s => selected.has(s.id)))
   const parentRef = useRef<HTMLDivElement>(null)
   const rowVirtualizer = useVirtualizer({
     count: hasNextPage ? items.length + 1 : items.length,
@@ -212,18 +93,25 @@ export function ConfigSessions() {
     if (areAllFilteredSelected) {
       setSelected(new Set())
     } else {
-      // 未加载完时需先拉全量再全选，否则仅选中已加载的 500/1000 条会导致“删不完”
+      // 未加载完时需分片拉全量再全选，后端 clamp 5000，>5000 时需循环
       if (hasNextPage) {
         try {
-          const all = await invoke<{ total: number, items: SessionFileInfo[] }>('get_session_files_paged', {
-            filter,
-            search: debouncedSearch || null,
-            sortKey,
-            sortAsc,
-            offset: 0,
-            limit: total,
-          })
-          setSelected(new Set(all.items.map(s => s.id)))
+          const chunk = 2000
+          const ids: string[] = []
+          for (let off = 0; off < total; off += chunk) {
+            const page = await invoke<{ total: number, items: SessionFileInfo[] }>('get_session_files_paged', {
+              filter,
+              search: debouncedSearch || null,
+              sortKey,
+              sortAsc,
+              offset: off,
+              limit: Math.min(chunk, total - off),
+            })
+            for (const s of page.items) ids.push(s.id)
+            if (page.items.length < chunk) break
+          }
+          if (ids.length > 0) setSelected(new Set(ids))
+          else setSelected(new Set(items.map(s => s.id)))
         } catch {
           // 兜底：至少选中已加载的
           setSelected(new Set(items.map(s => s.id)))
@@ -235,15 +123,7 @@ export function ConfigSessions() {
   }
   async function handleDelete(ids: string[]) {
     if (ids.length === 0) return
-    function truncatePreview(text: string, max = 24) {
-      const tt = text.trim()
-      return tt.length > max ? `${tt.slice(0, max)}…` : tt
-    }
-    const rawPreview = ids.slice(0, 3).map(id => truncatePreview(items.find(s => s.id === id)?.title || sessionsFallbackTitle(id)))
-    const preview = rawPreview.join('、')
-    const more = ids.length > 3 ? t('sessions.delete.batch_more', { count: ids.length - 3 }) : ''
-    const previewText = preview + more
-    const previewCapped = previewText.length > 72 ? `${previewText.slice(0, 72)}…` : previewText
+    const previewCapped = buildPreviewCapped(ids, items, t as (k: string, o?: unknown) => string)
     const confirmed = await openDialog({
       status: 'danger',
       title: t('sessions.delete.confirm_title'),
@@ -272,15 +152,7 @@ export function ConfigSessions() {
 
   async function handleRestore(ids: string[]) {
     if (ids.length === 0) return
-    function truncatePreview(text: string, max = 24) {
-      const tt = text.trim()
-      return tt.length > max ? `${tt.slice(0, max)}…` : tt
-    }
-    const rawPreview = ids.slice(0, 3).map(id => truncatePreview(items.find(s => s.id === id)?.title || id))
-    const preview = rawPreview.join('、')
-    const more = ids.length > 3 ? t('sessions.delete.batch_more', { count: ids.length - 3 }) : ''
-    const previewText = preview + more
-    const previewCapped = previewText.length > 72 ? `${previewText.slice(0, 72)}…` : previewText
+    const previewCapped = buildPreviewCapped(ids, items, t as (k: string, o?: unknown) => string)
     const confirmed = await openDialog({
       status: 'accent',
       title: t('sessions.restore.confirm_title'),
@@ -301,10 +173,6 @@ export function ConfigSessions() {
     } catch (e) {
       toast(String(e), { timeout: 3000 })
     }
-  }
-
-  function sessionsFallbackTitle(id: string) {
-    return id
   }
   async function handleOpenDir(id: string) {
     try {
